--- packages/devtools_shared/lib/src/url_utils.dart
@@ -1,29 +1,0 @@
-// Copyright 2019 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-/// Returns a normalized vm service uri.
-///
-/// Removes trailing characters, trailing url fragments, and decodes urls that
-/// were accidentally encoded.
-///
-/// For example, given a [value] of http://127.0.0.1:60667/72K34Xmq0X0=/#/vm,
-/// this method will return the URI http://127.0.0.1:60667/72K34Xmq0X0=/.
-///
-/// Returns null if the [Uri] parsed from [value] is not [Uri.absolute]
-/// (ie, it has no scheme or it has a fragment).
-Uri normalizeVmServiceUri(String value) {
-  value = value.trim();
-
-  // Cleanup encoded urls likely copied from the uri of an existing running
-  // DevTools app.
-  if (value.contains('%3A%2F%2F')) {
-    value = Uri.decodeFull(value);
-  }
-  final uri = Uri.parse(value.trim()).removeFragment();
-  if (!uri.isAbsolute) {
-    return null;
-  }
-  if (uri.path.endsWith('/')) return uri;
-  return uri.replace(path: uri.path);
-}
