+++ packages/devtools_shared/pubspec.yaml
@@ -9,6 +9,3 @@ homepage: https://github.com/flutter/devtools
 
 environment:
   sdk: '>=2.3.0 <3.0.0'
-
-dev_dependencies:
-  test: any
