+++ tool/travis.sh
@@ -99,12 +99,6 @@ if [ "$BOT" = "main" ]; then
     # Ensure we can build the app.
     flutter pub run build_runner build -o web:build --release
 
-    # Run shared tests.
-    pushd ../devtools_shared
-    dart pub get
-    dart test
-    popd
-
 elif [ "$BOT" = "test_ddc" ]; then
 
     flutter pub get
